Refactor: rename `protocol::wire_types` to `protocol::types`
This file doesn't just hold wire types, so the more generic name feels
better.
